chore: replace karma with wpt runner in user-interaction
