+++ plugins/web/opentelemetry-instrumentation-document-load/package.json
@@ -52,16 +52,11 @@
   },
   "devDependencies": {
     "@babel/core": "7.22.17",
-    "@jsdevtools/coverage-istanbul-loader": "3.0.5",
     "@opentelemetry/api": "^1.3.0",
-    "@rollup/plugin-commonjs": "^25.0.7",
-    "@rollup/plugin-node-resolve": "^15.2.3",
     "@types/chai": "^4.3.10",
     "@types/mocha": "8.2.3",
     "@types/node": "18.6.5",
     "@types/sinon": "10.0.18",
-    "@web/dev-server-esbuild": "^1.0.1",
-    "@web/dev-server-rollup": "^0.6.1",
     "@web/test-runner": "^0.18.0",
     "chai": "^4.3.10",
     "sinon": "15.2.0",
