+++ plugins/web/opentelemetry-instrumentation-user-interaction/package.json
@@ -16,8 +16,8 @@
     "version:update": "node ../../../scripts/version-update.js",
     "compile": "npm run version:update && tsc --build tsconfig.json tsconfig.esm.json tsconfig.esnext.json",
     "prepublishOnly": "npm run compile",
-    "tdd": "karma start",
-    "test:browser": "nyc karma start --single-run",
+    "tdd": "wtr --watch",
+    "test:browser": "wtr --coverage",
     "watch": "tsc --build --watch tsconfig.json tsconfig.esm.json tsconfig.esnext.json"
   },
   "keywords": [
@@ -49,34 +49,19 @@
   },
   "devDependencies": {
     "@babel/core": "7.22.17",
-    "@jsdevtools/coverage-istanbul-loader": "3.0.5",
     "@opentelemetry/api": "^1.3.0",
     "@opentelemetry/context-zone-peer-dep": "^1.8.0",
     "@opentelemetry/instrumentation-xml-http-request": "^0.49.1",
     "@opentelemetry/sdk-trace-base": "^1.8.0",
+    "@types/chai": "^4.3.10",
     "@types/jquery": "3.5.20",
-    "@types/mocha": "7.0.2",
+    "@types/mocha": "8.2.3",
     "@types/node": "18.6.5",
     "@types/sinon": "10.0.18",
-    "@types/webpack-env": "1.16.2",
-    "babel-loader": "8.2.2",
-    "karma": "6.3.16",
-    "karma-chrome-launcher": "^3.1.1",
-    "karma-coverage-istanbul-reporter": "3.0.3",
-    "karma-jquery": "0.2.4",
-    "karma-mocha": "2.0.1",
-    "karma-spec-reporter": "0.0.36",
-    "karma-webpack": "4.0.2",
-    "mocha": "7.2.0",
-    "nyc": "15.1.0",
-    "rimraf": "5.0.5",
+    "@web/test-runner": "^0.18.0",
+    "chai": "^4.3.10",
     "sinon": "15.2.0",
-    "ts-loader": "8.3.0",
-    "ts-mocha": "10.0.0",
     "typescript": "4.4.4",
-    "webpack": "4.46.0",
-    "webpack-cli": "4.7.2",
-    "webpack-merge": "5.8.0",
     "zone.js": "^0.11.4 || ^0.13.0 || ^0.14.0"
   },
   "dependencies": {
