+++ plugins/web/opentelemetry-instrumentation-user-interaction/test/userInteraction.test.ts
@@ -23,7 +23,8 @@ import {
 import { XMLHttpRequestInstrumentation } from '@opentelemetry/instrumentation-xml-http-request';
 import * as tracing from '@opentelemetry/sdk-trace-base';
 import { WebTracerProvider } from '@opentelemetry/sdk-trace-web';
-import * as assert from 'assert';
+// @ts-expect-error: not an export, but we want the prebundled version
+import chai from 'chai/chai.js';
 import * as sinon from 'sinon';
 import 'zone.js';
 import { UserInteractionInstrumentation } from '../src';
@@ -39,6 +40,7 @@ import {
   getData,
 } from './helper.test';
 
+const { assert } = chai as typeof import('chai');
 const FILE_URL =
   'https://raw.githubusercontent.com/open-telemetry/opentelemetry-js/main/package.json';
